4.21 done

+++ osa4/blogilista/controllers/blogs.js
@@ -4,14 +4,6 @@ const User = require('../models/user');
 const logger = require('../utils/logger');
 const jwt = require('jsonwebtoken');
 
-// ...
-const getTokenFrom = req => {
-  const authorization = req.get('authorization');
-  if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
-    return authorization.substring(7);
-  }
-  return null;
-};
 // show all blogs
 blogsRouter.get('/', async (req, res) => {
   const blog = await Blog
@@ -26,9 +18,8 @@ blogsRouter.get('/', async (req, res) => {
 // add a blog
 blogsRouter.post('/', async (req, res) => {
   const body = req.body;
-  const token = getTokenFrom(req);
-  const decodedToken = jwt.verify(token, process.env.SECRET);
-  if (!token || !decodedToken.id) {
+  const decodedToken = jwt.verify(req.token, process.env.SECRET);
+  if (!req.token || !decodedToken.id) {
     return res.status(401).json({ error: 'token missing or invalid' });
   }
   const user = await User.findById(decodedToken.id);
@@ -46,23 +37,7 @@ blogsRouter.post('/', async (req, res) => {
   await user.save();
   res.json(savedBlog.toJSON());
 });
-/*
-notesRouter.post('/', async (request, response) => {
-  const body = request.body
-  const token = getTokenFrom(request)
-  const decodedToken = jwt.verify(token, process.env.SECRET)
-  if (!token || !decodedToken.id) {
-    return response.status(401).json({ error: 'token missing or invalid' })
-  }
-  const user = await User.findById(decodedToken.id)
 
-  const note = new Note({
-    content: body.content,
-    important: body.important === undefined ? false : body.important,
-    date: new Date(),
-    user: user._id
-  })
-*/
 // show a blog with certain id
 blogsRouter.get('/:id', async (req, res) => {
   const blog = await Blog.findById(req.params.id);
@@ -75,8 +50,14 @@ blogsRouter.get('/:id', async (req, res) => {
 
 // delete blog
 blogsRouter.delete('/:id', async (req, res) => {
-  await Blog.findByIdAndRemove(req.params.id);
-  res.status(204).end();
+  const decodedToken = jwt.verify(req.token, process.env.SECRET);
+  const blog = await Blog.findById(req.params.id);
+  if (blog.user.toString() === decodedToken.id) {
+    await Blog.findByIdAndRemove(req.params.id);
+    res.status(204).end();
+  } else {
+    return res.status(401).json({ error: 'not authorized to delete' });
+  }
 });
 
 //modificate blog
